requirements.txt
playwright
icalendar
python-dateutil